pyproject.toml
[tool.poetry]
name = "backend-artistmanagementtsystem"
version = "0.1.0"
description = ""
authors = ["phuyalbish <phuyalbish@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
django-environ = "^0.11.2"
django-cors-headers = "^4.3.1"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
mysqlclient = "^2.2.7"
whitenoise = "^6.8.2"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pylint = "^3.3.6"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
